improvements on resolve-tree-parser
