+++ packages/graphql/src/api-v6/queryIRFactory/resolve-tree-parser/parse-global-resolve-info-tree.ts
@@ -1,0 +1,66 @@
+/*
+ * Copyright (c) "Neo4j"
+ * Neo4j Sweden AB [http://neo4j.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import type { ResolveTree } from "graphql-parse-resolve-info";
+import type { ConcreteEntity } from "../../../schema-model/entity/ConcreteEntity";
+import type { GraphQLTree } from "./graphql-tree/graphql-tree";
+import { parseNode } from "./parse-node";
+
+/** Parses the resolve info tree for a global node query */
+export function parseGlobalNodeResolveInfoTree({
+    resolveTree,
+    entity,
+}: {
+    resolveTree: ResolveTree;
+    entity: ConcreteEntity;
+}): GraphQLTree {
+    const entityTypes = entity.typeNames;
+    resolveTree.fieldsByTypeName[entityTypes.node] = {
+        ...resolveTree.fieldsByTypeName["Node"],
+        ...resolveTree.fieldsByTypeName[entityTypes.node],
+    };
+    const node = resolveTree ? parseNode(resolveTree, entity) : undefined;
+
+    return {
+        alias: resolveTree.alias,
+        args: {
+            where: {
+                node: {
+                    id: { equals: resolveTree.args.id as any },
+                },
+            },
+        },
+        name: resolveTree.name,
+        fields: {
+            connection: {
+                alias: "connection",
+                args: {},
+                fields: {
+                    edges: {
+                        alias: "edges",
+                        args: {},
+                        fields: {
+                            node,
+                        },
+                    },
+                },
+            },
+        },
+    };
+}
