+++ packages/graphql/tests/tck/directives/alias.test.ts
@@ -48,81 +48,6 @@ describe("Cypher alias directive", () => {
         });
     });
 
-    test("Simple relation", async () => {
-        const query = /* GraphQL */ `
-            {
-                actors {
-                    name
-                    city
-                    actedIn {
-                        title
-                        rating
-                    }
-                }
-            }
-        `;
-
-        const result = await translateQuery(neoSchema, query);
-
-        expect(formatCypher(result.cypher)).toMatchInlineSnapshot(`
-            "MATCH (this:Actor)
-            CALL {
-                WITH this
-                MATCH (this)-[this0:ACTED_IN]->(this1:Movie)
-                WITH this1 { .title, rating: this1.ratingPropInDb } AS this1
-                RETURN collect(this1) AS var2
-            }
-            RETURN this { .name, city: this.cityPropInDb, actedIn: var2 } AS this"
-        `);
-
-        expect(formatParams(result.params)).toMatchInlineSnapshot(`"{}"`);
-    });
-
-    test("With relationship properties", async () => {
-        const query = /* GraphQL */ `
-            {
-                actors {
-                    name
-                    city
-                    actedInConnection {
-                        edges {
-                            properties {
-                                character
-                                screenTime
-                            }
-                            node {
-                                title
-                                rating
-                            }
-                        }
-                    }
-                }
-            }
-        `;
-
-        const result = await translateQuery(neoSchema, query);
-
-        expect(formatCypher(result.cypher)).toMatchInlineSnapshot(`
-            "MATCH (this:Actor)
-            CALL {
-                WITH this
-                MATCH (this)-[this0:ACTED_IN]->(this1:Movie)
-                WITH collect({ node: this1, relationship: this0 }) AS edges
-                WITH edges, size(edges) AS totalCount
-                CALL {
-                    WITH edges
-                    UNWIND edges AS edge
-                    WITH edge.node AS this1, edge.relationship AS this0
-                    RETURN collect({ properties: { character: this0.characterPropInDb, screenTime: this0.screenTime, __resolveType: \\"ActorActedInProps\\" }, node: { title: this1.title, rating: this1.ratingPropInDb, __resolveType: \\"Movie\\" } }) AS var2
-                }
-                RETURN { edges: var2, totalCount: totalCount } AS var3
-            }
-            RETURN this { .name, city: this.cityPropInDb, actedInConnection: var3 } AS this"
-        `);
-
-        expect(formatParams(result.params)).toMatchInlineSnapshot(`"{}"`);
-    });
-
     test("Create mutation", async () => {
         const query = /* GraphQL */ `
             mutation {
